+++ apps/wsb-task/src/commands/artifacts/copy_as_artifact/params.rs
@@ -1,4 +1,4 @@
-use crate::commands::artifacts::artifacts_dir;
+use crate::core::env::artifacts_dir;
 use crate::core::targets::{AsBuildTarget, BuildTarget};
 use std::path::{Path, PathBuf};
