+++ apps/wsb-task/src/commands/mod.rs
@@ -1,3 +1,2 @@
 mod artifacts;
 pub use artifacts::copy_as_artifact;
-pub use artifacts::{app_names, artifacts_dir, executable_names};
