+++ apps/wsb-task/src/core/mod.rs
@@ -1,6 +1,8 @@
 mod action;
 pub use action::ActionOutput;
 
+pub mod env;
+
 pub mod support;
 
 #[macro_use]
